requirements.txt
Django==3.0.2
djangorestframework==3.11.0
requests==2.22.0
